Added ResourceRetrivalException exception class to handle exception when retrieving data

+++ manager/src/main/java/com/ldtech/services/impl/EmployeeServiceImpl.java
@@ -3,6 +3,7 @@ package com.ldtech.services.impl;
 import com.ldtech.entities.Employee;
 import com.ldtech.entities.Role;
 import com.ldtech.exceptions.ResourceNotFoundException;
+import com.ldtech.exceptions.ResourceRetrivalException;
 import com.ldtech.repositories.EmployeeRepository;
 import com.ldtech.services.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,11 +29,24 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     public List<Employee> getAllEmployees() {
-        return employeeRepository.findAll();
+        try {
+            return employeeRepository.findAll();
+        } catch (Exception e) {
+            throw new ResourceRetrivalException("Failed to retrieve all employees", e);
+        }
     }
 
     @Override
     public void deleteEmployee(String employeeId) {
+        // Check if the employee exists
+        boolean exists = employeeRepository.existsById(employeeId);
+        if (!exists) {
+            // Throw custom exception if employee does not exist
+            throw new ResourceNotFoundException("Employee", "ID", employeeId);
+        }
+
+        // Proceed to delete the employee if they exist
         employeeRepository.deleteById(employeeId);
     }
+
 }
